Scripts/requirements.txt
numpy
pandas
matplotlib
#tensorflow
scikit-learn
Pillow
seaborn
